scripts/requirements.txt
click==8.1.7
feedparser==6.0.11
joblib==1.4.2
nltk==3.9.1
python-dateutil==2.9.0.post0
regex==2024.11.6
sgmllib3k==1.0.0
six==1.16.0
textblob==0.18.0.post0
tqdm==4.67.1
